refactor!: remove multifile_strategy (this is now supported by Neovim natively)
This is not technically a breaking change because the plugin
requires Neovim 0.8 which always refers to the latest nightly version.
However, when renaming across multiple buffers (which should not happen
too often honestly) with a lower nightly version, highlights won't
be cleared anymore.
Make sure to upgrade your Neovim version to the latest available nightly
version for the best experience.