fix: only exit the rename dialog when all clients have been exhausted
The rename dialog may close prematurely when multiple LSPs are capable
of modifying the current buffer. This typically happens when more than
one LSP is attached to the same buffer, with each handling different
parts of the file. It's a common scenario in multi-language files—for
instance, VueJS files, where the <template> block is handled by Volar
and the <script> block by tsserver (ts_ls).